One Storybook entry for the top bar, not two
